Panier fini

+++ shop/view/page/basket/basket.php
@@ -6,7 +6,7 @@
 
 		$total = 0;
 
-		if(isset($_SESSION["basket"])){
+		if(isset($_SESSION["basket"]) && isset($products)){
 			echo '<table>';
 			echo '<tr>';
 			echo '<th>Description</th>';
@@ -23,26 +23,26 @@
 				<td> CHF '. $_SESSION["basket"][$product[0]['idProduct']] * $product[0]['proPrice'] . '</td>';
 				?>
 				<td>
-				<a class="btn btn-default" href="index.php?controller=basket&action=delete&id=<?= $product[0]['idProduct'] ?>">
-					<span class="glyphicon glyphicon-pencil" aria-hidden="true"></span></a>
-				<a onclick="return confirm(\'Etes-vous sûr ? \')" class="btn btn-default"
-				href="index.php?controller=basket&action=delete&id=<?= $product[0]['idProduct'] ?>">
-					<span class="glyphicon glyphicon-remove" aria-hidden="true"></span></a>
-			</td>
-				<?=
+					<a class="btn btn-default" href="index.php?controller=basket&action=add&id=<?= $product[0]['idProduct'] ?>&quant=<?= $product[0]['proQuantity'] ?>">+</a>
+					<a class="btn btn-default" href="index.php?controller=basket&action=remove&id=<?= $product[0]['idProduct'] ?>">-</a>
+					<a onclick="return confirm('Etes-vous sûr ?')" class="btn btn-default" href="index.php?controller=basket&action=delete&id=<?= $product[0]['idProduct'] ?>">
+						<span class="glyphicon glyphicon-remove" aria-hidden="true"></span></a>
+				</td>
+				<?php
 				$total += $_SESSION["basket"][$product[0]['idProduct']] * $product[0]['proPrice'];
 			}
 			echo '<tr>
 			<td> Total </td>
 			<td> </td>
 			<td> </td>
-			<td> CHF '. $total .'</td>';
+			<td> CHF '. $total .'</td>
+			<td> </td>';
 			echo '</table>';
+			echo'<a class="btn btn-default command" href="index.php?controller=order&action=delivery">Passer commande</a>';
 		}
 		else{
-			echo "T'es pauvre";
+			echo "Le panier est actuellement vide";
 		}
 		?>
-
 	</div>
 </div>
